more migrations for images
